[FIXED] moved click event to JS instead of HTML
