+++ .gitignore
@@ -108,3 +108,4 @@ dist
 webpack.config.js
 package-lock.json
 package.json
+babel.config.json
