+++ src/dropzone.js
