+++ src/slider.js
@@ -1,10 +1,15 @@
+/* Slider */
+
 import slideImage from "C:/wamp64/www/team-four-webapp/src/assets/images/slider-img.png";
 import slideImageTwo from "C:/wamp64/www/team-four-webapp/src/assets/images/open-graph-preview.png";
 
-/* Slider */
-
+const prevButton = document.querySelector(".prev");
+const nextButton = document.querySelector(".next");
 var slideIndex = 1;
 showSlides(slideIndex);
+
+prevButton.addEventListener("click", plusSlides(slideIndex-2));
+nextButton.addEventListener("click", plusSlides(slideIndex));
 
 // Next/previous controls
 function plusSlides(n) {
